Introduce focus blocks
Focus blocks is an attempt at offloading the overhead of keeping track
of what to be focusing on. Define blocks in the evening or the early
morning, and then use them as a guide throughout the day.

The implementation is a bit hacky, but I think it should be good enough.